pkg/ast_core/ast_map_extractor.go
package ast_core

import (
	"github.com/KoNekoD/go-deptrac/pkg/ast_core/ast_map"
	"github.com/KoNekoD/go-deptrac/pkg/input_collector_core"
)

type AstMapExtractor struct {
	inputCollector input_collector_core.InputCollectorInterface
	astLoader      *AstLoader
	astMapCache    *ast_map.AstMap
}

func NewAstMapExtractor(inputCollector input_collector_core.InputCollectorInterface, astLoader *AstLoader) *AstMapExtractor {
	return &AstMapExtractor{
		inputCollector: inputCollector,
		astLoader:      astLoader,
		astMapCache:    nil,
	}
}

func (e *AstMapExtractor) Extract() (*ast_map.AstMap, error) {
	if e.astMapCache == nil {
		collected, err := e.inputCollector.Collect()

		if err != nil {
			return nil, err
		}

		createdAstMap, errCreateAstMap := e.astLoader.CreateAstMap(collected)
		if errCreateAstMap != nil {
			return nil, errCreateAstMap
		}

		e.astMapCache = createdAstMap
	}

	return e.astMapCache, nil
}
